requirements.txt
icecream==2.1.3
paho-mqtt==2.0.0 
pendulum==3.0.0
influxdb-client==1.40.0